requirements.txt
asgiref==3.4.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.1.0
django-formtools==2.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-taggit==1.5.1
django-tinymce==3.3.0
gunicorn==20.1.0
importlib-metadata==4.8.2
jaconv==0.3
Pillow==8.4.0
psycopg2==2.9.1
pykakasi==2.2.1
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.6.0
